Crate pallet_xyk_liquidity_mining

Overview

This pallet provides functionality for liquidity mining programs with time incentive(loyalty factor) and multiple incentives scheme for XYK pools. Users are rewarded for each period they stay in liq. mining program.

Reward per one period is derived from the user’s loyalty factor which grows with time(periods) the user is in the liq. mining and amount of LP shares user locked into deposit. User’s loyalty factor is reset if the user exits and reenters liquidity mining. User can claim rewards without resetting loyalty factor, only withdrawing shares is penalized by loyalty factor reset. User is rewarded from the next period after he enters.

Multiple Incentives

This feature allow users to redeposit already deposited LP shares to multiple yield farms and receive incentives from this farms. LP shares can be redeposited only to different yield farms running liquidity mining for same pair of assets.

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_xyk_liquidity_mining

Type Definitions

PeriodOf 🔒